Use $pty->getline and $pty->read as loop terminator instead of $pty->is_active
If the command writes some data, waits a bit, writes some more data and then
exits quickly then not all of the output will be read because a small amount
of data will be read in the first loop while the process is still running,
then the process will have exited and the loop will terminate,
leaving the remaining output unread.

Using the read data as loop terminator ensures that it will all be read.

Found-by: diff -u <(pipetty mr -m status) <(pipetty mr -m status)