Handle empty config item values by using the empty string.
split() on an empty string produces an empty array and
shift() on an empty array produces an undefined value,
which causes Perl to warn when the undefined value is
concatenated with the field name and the end of line.

$ mr config foo bar=abs
Use of uninitialized value in concatenation (.) or string at mr line 1839.

Fixes: commit c934687bf61031c926d525f0149d4d4fe324269d