sort related software alphabetically
It seems we weren't really sorting that in any meaningful
way. It *seems* we were incrementally adding new projects as we were
finding them, but that seems like a rather incoherent approach,
because it doesn't really reflect the value of any of those projects,
or even when they were created (or updated?) last, which would be more
relevant information.